+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## 1.0.3 (2014-04-22)
+* added missed sandbox key to UrlBuilder for sandbox-secure.xsolla.com
+* fixed wrong error code for IPN requests with zero valued parameters
+
 ## 1.0.2 (2014-02-27)
 * add `$baseUrl` optional argument to `UrlBuilder::getUrl()` and `UrlBuilder::SANDBOX_URL` constant